Revert "no message"
This reverts commit c0cad07d841f8ed29607af8ffc7a572d245c7d86.

+++ Svelto.ECS/DynamicEntityDescriptor.cs
@@ -44,7 +44,7 @@ namespace Svelto.ECS
             var extraEntities = extraEntityBuilders.ToArrayFast();
             var extraEntitiesLength = extraEntityBuilders.Count;
 
-            _entitiesToBuild = Construct((int) extraEntitiesLength, extraEntities,
+            _entitiesToBuild = Construct(extraEntitiesLength, extraEntities,
                 EntityDescriptorTemplate<TType>.descriptor.entitiesToBuild);
         }
